Fixed bug when destination is outside gtd dir
org-gtd-project-extend--apply properly builds a list of potential
targets from the agenda + gtd files.

There does not seem to be a way to differentiate between two
headings (projects) in different files with the same name

Once the function asks the user for the destination, it then
searches the gtd directory for such header.  If the project is
NOT in the gtd directory, it fails with an error.

This patch improves the functionality by:

1. Keep a pair of buffer/project-name from agenda + gtd files, instead of just the headers.
2. Ask user for destination with both buffer name and header
3. Use the buffer instead of searching again in the gtd directory.